+++ txnbuild/CHANGELOG.md
@@ -6,7 +6,9 @@ file.  This project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 
 * Add support for getting the hex-encoded transaction hash with `Transaction.HashHex` method.
-* Add support for getting the `TransactionEnvelope` struct with `Transaction.TxEnvelope` method.
+* Add support for getting the `TransactionEnvelope` struct with `Transaction.TxEnvelope` method ([#1415](https://github.com/stellar/go/issues/1415)).
+* `AllowTrust` operations no longer requires the asset issuer, only asset code is required ([#1330](https://github.com/stellar/go/issues/1330)).
+
 
 ## [v1.2.0](https://github.com/stellar/go/releases/tag/horizonclient-v1.2.0) - 2019-05-16